Refactor component file hierarchy and imports for better organization
