+++ src/components/generic/BlurText.tsx
