+++ src/components/generic/line-chart-divider/LineChartDividerPeriod.tsx
