+++ src/components/projects/cafeteria-manager/cash-register/FakeCashRegister.tsx
