+++ src/components/projects/cafeteria-manager/stats/FakeStats.tsx
